+++ contracts/cosmwasm-vm/cw-common/src/lib.rs
@@ -11,6 +11,7 @@ pub mod query_helpers;
 pub mod raw_types;
 pub mod types;
 pub mod utils;
+pub mod xcall_connection_msg;
 
 use bech32::FromBase32;
 use cosmwasm_std::{from_binary, Addr, Binary, Deps, StdError};
